+++ src/main/java/com/mraof/minestuck/alchemy/recipe/generator/recipe/InterpreterSerializer.java
@@ -1,0 +1,10 @@
+package com.mraof.minestuck.alchemy.recipe.generator.recipe;
+
+import com.google.gson.JsonElement;
+
+public abstract class InterpreterSerializer<T extends RecipeInterpreter>
+{
+	public abstract T read(JsonElement json);
+	
+	public abstract JsonElement write(T interpreter);
+}