+++ src/main/java/com/mraof/minestuck/client/gui/AlchemiterScreen.java
@@ -85,8 +85,7 @@ public class AlchemiterScreen extends Screen
 		
 		
 		//Calculate the grist set
-		GristSet set;
-		set = alchemiter.getGristCost(itemQuantity);
+		GristSet set = alchemiter.getGristCost(itemQuantity);
 		//draw the grist board	//TODO Handle select mode correctly
 		GuiUtil.drawGristBoard(poseStack, set, AlchemyHelper.getDecodedItem(alchemiter.getDowel()).getItem() == MSItems.CAPTCHA_CARD.get() ? GuiUtil.GristboardMode.LARGE_ALCHEMITER_SELECT : GuiUtil.GristboardMode.LARGE_ALCHEMITER, (width - guiWidth) / 2 + 88, (height - guiHeight) / 2 + 13, font);
 		//draw the grist