+++ src/main/java/com/mraof/minestuck/client/gui/toasts/GristToast.java
@@ -236,11 +236,11 @@ public class GristToast implements Toast
 		GristHelper.EnumSource source = packet.source();
 		ToastComponent toasts = Minecraft.getInstance().getToasts();
 		
-		for(GristAmount pair : packet.gristValue().getAmounts())
+		for(GristAmount pair : packet.gristValue().asAmounts())
 		{
 			//the pair has to be split into two new variables because Map.Entry is immutable.
-			GristType type = pair.getType();
-			long difference = pair.getAmount();
+			GristType type = pair.type();
+			long difference = pair.amount();
 			long total = cache.set().getGrist(type);
 			
 			if(difference == 0)